use a generic way to test methods of PD Client (#2988)
